Add files via upload
edited and linked esma's pages (take action and about us) to all the pages, made some changes so that our code is more modular. Moved some pages around so the directory is more organized and makes sense. Changed header, footer and sidebar css files so it works with the private pages too. Gonna upload on my fork and submit a pull request.This should be the finalized look for our pages. All we need to do is figure out the login/registration and passage to dashboard now