Final Style Commit
edited and linked esma's pages (take action and about us) to all the pages, made some changes so that our code is more modular. Moved some pages around so the directory is more organized and makes sense. Changed header, footer and sidebar css files so it works with the private pages too.